+++ projects/CoinFlip/CoinFlip.Model2/src/Misc/IValueTotalAvail.cs
@@ -5,12 +5,12 @@ namespace CoinFlip
 	public interface IValueTotalAvail
 	{
 		/// <summary>Value of the coin (probably in USD)</summary>
-		decimal Value { get; }
+		double Value { get; }
 
 		/// <summary>The total amount of the coin (in coin currency)</summary>
-		Unit<decimal> Total { get; }
+		Unit<double> Total { get; }
 
 		/// <summary>The available amount of the coin (in coin currency)</summary>
-		Unit<decimal> Available { get; }
+		Unit<double> Available { get; }
 	}
 }
